# IEnumerable.Shuffle performance enhancement