--- a/Spk.Common.sln
+++ b/Spk.Common.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26730.10
+VisualStudioVersion = 15.0.26730.8
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{8D4155F3-D442-4D9D-AA51-A600AF6E5C40}"
 EndProject
@@ -23,7 +23,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Spk.Common.Tests.Helpers", 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Spk.Common.Tests.Validators", "test\Spk.Common.Tests.Validators\Spk.Common.Tests.Validators.csproj", "{0791E192-E12E-457F-998E-40B9EFF225EF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Spk.Common.Benchmarks", "test\Spk.Common.Benchmarks\Spk.Common.Benchmarks.csproj", "{FC9ACD5D-FC34-46F6-A585-F9AB1811AF71}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Spk.Common.Benchmarks", "test\Spk.Common.Benchmarks\Spk.Common.Benchmarks.csproj", "{33ACE061-8831-45D8-9864-386627E9265B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -47,10 +47,10 @@ Global
 		{0791E192-E12E-457F-998E-40B9EFF225EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0791E192-E12E-457F-998E-40B9EFF225EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0791E192-E12E-457F-998E-40B9EFF225EF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FC9ACD5D-FC34-46F6-A585-F9AB1811AF71}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FC9ACD5D-FC34-46F6-A585-F9AB1811AF71}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FC9ACD5D-FC34-46F6-A585-F9AB1811AF71}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FC9ACD5D-FC34-46F6-A585-F9AB1811AF71}.Release|Any CPU.Build.0 = Release|Any CPU
+		{33ACE061-8831-45D8-9864-386627E9265B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{33ACE061-8831-45D8-9864-386627E9265B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{33ACE061-8831-45D8-9864-386627E9265B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{33ACE061-8831-45D8-9864-386627E9265B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -60,7 +60,7 @@ Global
 		{5B3C8276-94D3-4297-9662-B02F133F5C67} = {354A7D2B-5951-4ED8-A87B-2FC65D72092C}
 		{12F000C1-FF48-4099-8FB5-6A6BB4A6EB4D} = {8D4155F3-D442-4D9D-AA51-A600AF6E5C40}
 		{0791E192-E12E-457F-998E-40B9EFF225EF} = {8D4155F3-D442-4D9D-AA51-A600AF6E5C40}
-		{FC9ACD5D-FC34-46F6-A585-F9AB1811AF71} = {8D4155F3-D442-4D9D-AA51-A600AF6E5C40}
+		{33ACE061-8831-45D8-9864-386627E9265B} = {8D4155F3-D442-4D9D-AA51-A600AF6E5C40}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {2029332C-C6D7-4771-B403-A6E96843640E}