--- a/test/Spk.Common.Benchmarks/Program.cs
+++ b/test/Spk.Common.Benchmarks/Program.cs
@@ -3,9 +3,9 @@ using Spk.Common.Benchmarks.Helpers.IEnumerable;
 
 namespace Spk.Common.Benchmarks
 {
-    internal class Program
+    class Program
     {
-        private static void Main(string[] args)
+        static void Main(string[] args)
         {
             BenchmarkRunner.Run<ShuffleExtensionBenchmarks>();
         }