--- a/test/Spk.Common.Benchmarks/Spk.Common.Benchmarks.csproj
+++ b/test/Spk.Common.Benchmarks/Spk.Common.Benchmarks.csproj
@@ -133,6 +133,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Helpers\IEnumerable\ShuffleExtensionBenchmarks.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>