… test/Spk.Common.Benchmarks/App.config
… test/Spk.Common.Benchmarks/Properties/AssemblyInfo.cs
… test/Spk.Common.Benchmarks/packages.config