NOTE: Contains #8, please review this one before if not already done.


Before optimization:

``` ini

BenchmarkDotNet=v0.10.9, OS=Windows 10 Redstone 2 (10.0.15063)
Processor=Intel Core i5-5257U CPU 2.70GHz (Broadwell), ProcessorCount=4
Frequency=2636717 Hz, Resolution=379.2595 ns, Timer=TSC
  [Host]     : .NET Framework 4.7 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.2115.0  [AttachedDebugger]
  DefaultJob : .NET Framework 4.7 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.2115.0


```
 |            Method |     Mean |     Error |    StdDev |
 |------------------ |---------:|----------:|----------:|
 | Shuffle_Benchmark | 38.02 ns | 0.6542 ns | 0.6719 ns |



After optimization:

``` ini

BenchmarkDotNet=v0.10.9, OS=Windows 10 Redstone 2 (10.0.15063)
Processor=Intel Core i5-5257U CPU 2.70GHz (Broadwell), ProcessorCount=4
Frequency=2636717 Hz, Resolution=379.2595 ns, Timer=TSC
  [Host]     : .NET Framework 4.7 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.2115.0  [AttachedDebugger]
  DefaultJob : .NET Framework 4.7 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.2115.0


```
 |            Method |     Mean |     Error |   StdDev |
 |------------------ |---------:|----------:|---------:|
 | Shuffle_Benchmark | 25.13 ns | 0.5871 ns | 1.567 ns |