@thatfrankdev 2017-11-03
We had to make it target netcoreapp2.0 to fit with our libraries. 🚀